Build Your own Cryptocurrency Exchange with Binance Clone Script
A Binance clone script is a white-label software that is based on the Binance cryptocurrency exchange platform. This means that it includes all of the features and functionality of Binance, such as trading, buying, selling, order matching, and so on. Binance clone scripts are a popular option for businesses that want to launch their own cryptocurrency exchange platform quickly and cost-effectively.